+++ src/EntityInterface.php
@@ -18,9 +18,10 @@ interface EntityInterface
     /**
      * Create an entity from array.
      * @param string $_arr
+     * @param string[] $_array_keys The keys of sub array, from which we need to extract the entity.
      * @return EntityInterface|null
      */
-    public static function fromJson(string $_arr): ?EntityInterface;
+    public static function fromJson(string $_arr, string ...$_array_keys): ?EntityInterface;
 
     /**
      * Convert an entity to array.